requirements.txt
# Dependencias para Plataforma de Cámaras Trampa con IA

# Core
streamlit>=1.28.0
pandas>=2.0.0
openpyxl>=3.1.0
Pillow>=10.0.0

# Análisis de datos
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0

# IA y Deep Learning (GPU RTX/CUDA)
# IMPORTANTE: Instalar PyTorch con CUDA desde https://pytorch.org/get-started/locally/
# Para RTX con CUDA 11.8+:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
torch>=2.0.0
torchvision>=0.15.0

# Procesamiento de imágenes
opencv-python>=4.8.0

# Utilidades
tqdm>=4.66.0
requests>=2.31.0

# Visualización
matplotlib>=3.7.0
plotly>=5.17.0

# Base de datos
# sqlite3 viene incluido con Python

# Nota: Si NO tienes GPU CUDA, la plataforma funcionará en modo asistido manual
# sin necesidad de torch/torchvision. Puedes comentar esas líneas.
